tesseractをgitからcloneしてビルドできるようにした。

+++ tesseract/tesseract.sh
@@ -30,24 +30,31 @@ echo "tesseract is already installed."
 exit 0
 fi
 
-TESSERACT_VERSION=5.0.0-alpha-20210401
-TESSERACT_TAG=$TESSERACT_VERSION
-TESSERACT_ARCHIVE=tesseract-$TESSERACT_TAG.tar.gz
-TESSERACT_SRC_DIR=tesseract-$TESSERACT_VERSION
-TESSERACT_BUILD_DIR=$TESSERACT_SRC_DIR-$BIT
 
-if [ "$TESSERACT_VERSION" == "master" ]; then
-rm $TESSERACT_ARCHIVE 2> /dev/null
+if [ "$TESSERACT_GIT" != "" ]; then
+    TESSERACT_SRC_DIR=tesseract-git
+    git clone https://github.com/tesseract-ocr/tesseract.git $TESSERACT_SRC_DIR 2> /dev/null
+    pushd $TESSERACT_SRC_DIR
+    git pull
+else
+    TESSERACT_VERSION=5.0.0-alpha-20210401
+    TESSERACT_TAG=$TESSERACT_VERSION
+    TESSERACT_ARCHIVE=tesseract-$TESSERACT_TAG.tar.gz
+    TESSERACT_SRC_DIR=tesseract-$TESSERACT_VERSION
+    TESSERACT_BUILD_DIR=$TESSERACT_SRC_DIR-$BIT
+
+    if [ "$TESSERACT_VERSION" == "master" ]; then
+    rm $TESSERACT_ARCHIVE 2> /dev/null
+    fi
+
+    if [ ! -e $TESSERACT_ARCHIVE ]; then
+    wget -c https://github.com/tesseract-ocr/tesseract/archive/$TESSERACT_TAG/$TESSERACT_ARCHIVE
+    fi
+    rm -rf $TESSERACT_SRC_DIR $TESSERACT_BUILD_DIR 
+    tar xf $TESSERACT_ARCHIVE
+    mv $TESSERACT_SRC_DIR $TESSERACT_BUILD_DIR
+    pushd $TESSERACT_BUILD_DIR
 fi
-
-
-if [ ! -e $TESSERACT_ARCHIVE ]; then
-wget -c https://github.com/tesseract-ocr/tesseract/archive/$TESSERACT_TAG/$TESSERACT_ARCHIVE
-fi
-rm -rf $TESSERACT_SRC_DIR $TESSERACT_BUILD_DIR 
-tar xf $TESSERACT_ARCHIVE
-mv $TESSERACT_SRC_DIR $TESSERACT_BUILD_DIR
-pushd $TESSERACT_BUILD_DIR
 
 #asciidocが動かない問題への暫定対応
 sed -i -e 's/AM_CONDITIONAL(\[ASCIIDOC\], true)/AM_CONDITIONAL([ASCIIDOC], false)/' configure.ac
